Fix circular imports for util functions
