+++ mirar/data/utils/plot_image.py
