+++ mirar/processors/candidates/dataframe_writer.py
@@ -3,6 +3,7 @@ Module with classes to write a candidate table to a pandas dataframe
 """
 import logging
 import os
+from pathlib import Path
 from typing import Optional
 
 from mirar.data import SourceBatch
@@ -20,15 +21,19 @@ class DataframeWriter(BaseDataframeProcessor):
     base_key = "DFWRITE"
 
     def __init__(
-        self, output_dir_name: Optional[str] = None, output_dir: str = base_output_dir
+        self,
+        output_dir_name: Optional[str] = None,
+        output_dir: str | Path = base_output_dir,
     ):
         super().__init__()
         self.output_dir_name = output_dir_name
-        self.output_dir = output_dir
+        self.output_dir = Path(output_dir)
         logger.debug(f"Saving candidates to {self.output_dir_name}")
 
     def __str__(self) -> str:
-        return "Processor to save candidates as a pickle file. "
+        return (
+            f"Processor to save candidates to {self.output_dir_name} as a json file. "
+        )
 
     def _apply_to_candidates(
         self,
@@ -47,10 +52,9 @@ class DataframeWriter(BaseDataframeProcessor):
 
         for source_list in batch:
             candidate_table = source_list.get_data()
-
             df_basepath = os.path.basename(
                 candidate_table.loc[0]["diffimname"]
-            ).replace(".fits", ".candidates.pkl")
+            ).replace(".fits", ".candidates.json")
             df_path = get_output_path(
                 df_basepath,
                 dir_root=self.output_dir_name,
@@ -58,6 +62,6 @@ class DataframeWriter(BaseDataframeProcessor):
                 output_dir=self.output_dir,
             )
             logger.info(f"Writing dataframe to {df_path}")
-            candidate_table.to_pickle(df_path)
+            candidate_table.to_json(df_path)
 
         return batch
